# prover/docs/swagger.yaml
basePath: /
definitions:
  api.BatchResponse:
    properties:
      batchHash:
        type: string
      batchNumber:
        type: integer
      newStateRoot:
        type: string
      status:
        type: integer
      submitter:
        type: string
      timestamp:
        type: integer
      txHash:
        type: string
      txHashes:
        description: L2 transaction hashes
        items:
          type: string
        type: array
    type: object
  api.StatusResponse:
    properties:
      batch_number:
        type: integer
      note:
        type: string
      status:
        type: string
      timestamp:
        type: integer
    type: object
info:
  contact: {}
  description: API for the Nowa-ZK Rollup Prover
  title: Nowa-ZK Prover API
  version: "1.0"
paths:
  /:
    get:
      description: Redirects to Swagger UI
      responses:
        "302":
          description: Found
      summary: List available endpoints
      tags:
      - Meta
  /batches:
    get:
      description: Retrieve multiple batch metadata entries with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page - 10, 25, 50, or 100 (default: 25)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated batch list
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get batches with pagination
      tags:
      - Batches
  /batches/{id}:
    get:
      description: Get details of a specific batch from local storage
      parameters:
      - description: Batch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BatchResponse'
        "400":
          description: Invalid batch ID
          schema:
            type: string
        "404":
          description: Batch not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get batch by ID
      tags:
      - Batches
  /batches/latest:
    get:
      description: Get details of the latest verified batch from the contract
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BatchResponse'
        "404":
          description: No batches found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get latest verified batch
      tags:
      - Batches
  /contracts:
    get:
      description: Get a list of deployed smart contracts and their addresses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No contracts found
          schema:
            type: string
      summary: Get deployed smart contracts
      tags:
      - Meta
  /status/{id}:
    get:
      description: Get proof generation status for a batch
      parameters:
      - description: Batch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Invalid batch ID
          schema:
            type: string
      summary: Get proof status
      tags:
      - Proofs
swagger: "2.0"