# prover/internal/api/server.go
package api

import (
	"encoding/json"
	"fmt"
	"math/big"
	"strconv"

	"github.com/ethereum/go-ethereum/accounts/abi/bind"
	"github.com/ethereum/go-ethereum/common"
	"github.com/gofiber/fiber/v2"
	"github.com/gofiber/fiber/v2/middleware/cors"
	"github.com/gofiber/fiber/v2/middleware/logger"
	"github.com/gofiber/swagger"
	"github.com/nowafinance/nowa-zk/prover/bindings"
	_ "github.com/nowafinance/nowa-zk/prover/docs" // Swagger docs
	"github.com/nowafinance/nowa-zk/prover/internal/storage"
)

// APIServer serves the Prover API
type APIServer struct {
	app      *fiber.App
	registry    *bindings.BatchRegistry
	store       *storage.ProverStore
	deployments map[string]string
	port        int
}

// BatchResponse represents the batch data returned by API
type BatchResponse struct {
	BatchNumber  uint64   `json:"batchNumber"`
	BatchHash    string   `json:"batchHash"`
	NewStateRoot string   `json:"newStateRoot"`
	Submitter    string   `json:"submitter"`
	Timestamp    uint64   `json:"timestamp"`
	Status       uint8    `json:"status"`
	TxHash       string   `json:"txHash,omitempty"`
	TxHashes     []string `json:"txHashes,omitempty"` // L2 transaction hashes
}

// StatusResponse represents the proof status
type StatusResponse struct {
	Status      string `json:"status"`
	BatchNumber uint64 `json:"batch_number"`
	Timestamp   int64  `json:"timestamp,omitempty"`
	Note        string `json:"note,omitempty"`
}

// NewAPIServer creates a new API server
func NewAPIServer(registry *bindings.BatchRegistry, store *storage.ProverStore, deployments map[string]string, port int) *APIServer {
	app := fiber.New(fiber.Config{
		DisableStartupMessage: true,
		AppName:               "Nowa-ZK Prover",
		JSONEncoder:           json.Marshal,
		JSONDecoder:           json.Unmarshal,
	})

	// Middleware
	app.Use(logger.New())
	app.Use(cors.New(cors.Config{
		AllowOrigins: "*",
		AllowMethods: "GET,POST,HEAD,PUT,DELETE,PATCH,OPTIONS",
		AllowHeaders: "Origin,Content-Type,Accept,Authorization",
	}))

	return &APIServer{
		app:         app,
		registry:    registry,
		store:       store,
		deployments: deployments,
		port:        port,
	}
}

// Start starts the API server
// @title Nowa-ZK Prover API
// @version 1.0
// @description API for the Nowa-ZK Rollup Prover
// @BasePath /
func (api *APIServer) Start() error {
	// Swagger
	api.app.Get("/swagger/*", swagger.New(swagger.Config{
		Title:       "Nowa-ZK Prover API",
		CustomStyle: `.swagger-ui .topbar { display: none }`,
	}))

	// Root endpoint
	api.app.Get("/", api.handleRoot)

	// Batches
	api.app.Get("/batches", api.handleGetBatches)
	api.app.Get("/batches/latest", api.handleLatestBatch)
	api.app.Get("/batches/:id", api.handleGetBatch)

	// Status
	api.app.Get("/status/:id", api.handleGetStatus)

	// Contracts
	api.app.Get("/contracts", api.handleGetContracts)

	addr := fmt.Sprintf(":%d", api.port)
	fmt.Printf("🌍 Starting Prover API on %s\n", addr)
	fmt.Printf("📖 Swagger UI available at http://localhost%s/swagger/index.html\n", addr)

	return api.app.Listen(addr)
}

// handleRoot godoc
// @Summary List available endpoints
// @Description Redirects to Swagger UI
// @Tags Meta
// @Success 302
// @Router / [get]
func (api *APIServer) handleRoot(c *fiber.Ctx) error {
	return c.Redirect("/swagger/index.html")
}

// handleGetContracts godoc
// @Summary Get deployed smart contracts
// @Description Get a list of deployed smart contracts and their addresses
// @Tags Meta
// @Produce json
// @Success 200 {object} map[string]string
// @Failure 404 {string} string "No contracts found"
// @Router /contracts [get]
func (api *APIServer) handleGetContracts(c *fiber.Ctx) error {
	if api.deployments == nil || len(api.deployments) == 0 {
		return c.Status(fiber.StatusNotFound).SendString("No deployed contracts found")
	}
	return c.JSON(api.deployments)
}

// handleLatestBatch godoc
// @Summary Get latest verified batch
// @Description Get details of the latest verified batch from the contract
// @Tags Batches
// @Produce json
// @Success 200 {object} BatchResponse
// @Failure 404 {string} string "No batches found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /batches/latest [get]
func (api *APIServer) handleLatestBatch(c *fiber.Ctx) error {
	totalBatches, err := api.registry.TotalBatches(&bind.CallOpts{})
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).SendString(fmt.Sprintf("Failed to get total batches: %v", err))
	}

	if totalBatches.Cmp(big.NewInt(0)) == 0 {
		return c.Status(fiber.StatusNotFound).SendString("No batches found")
	}

	batch, err := api.registry.GetBatch(&bind.CallOpts{}, totalBatches)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).SendString(fmt.Sprintf("Failed to get batch: %v", err))
	}

	// Try to get tx hash and tx hashes from local storage
	txHash := ""
	var txHashes []string
	if proofData, err := api.store.GetProof(totalBatches.Uint64()); err == nil && proofData != nil {
		txHash = proofData.TxHash
		txHashes = proofData.TxHashes
	}

	resp := BatchResponse{
		BatchNumber:  totalBatches.Uint64(),
		BatchHash:    common.BytesToHash(batch.BatchHash[:]).Hex(),
		NewStateRoot: common.BytesToHash(batch.NewStateRoot[:]).Hex(),
		Submitter:    batch.Submitter.Hex(),
		Timestamp:    batch.Timestamp.Uint64(),
		Status:       batch.Status,
		TxHash:       txHash,
		TxHashes:     txHashes,
	}

	return c.JSON(resp)
}

// handleGetBatch godoc
// @Summary Get batch by ID
// @Description Get details of a specific batch from local storage
// @Tags Batches
// @Param id path int true "Batch ID"
// @Produce json
// @Success 200 {object} BatchResponse
// @Failure 400 {string} string "Invalid batch ID"
// @Failure 404 {string} string "Batch not found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /batches/{id} [get]
func (api *APIServer) handleGetBatch(c *fiber.Ctx) error {
	idStr := c.Params("id")
	batchNumber, err := strconv.ParseUint(idStr, 10, 64)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).SendString("Invalid batch ID")
	}

	// Query local storage (not blockchain contract)
	proofData, err := api.store.GetProof(batchNumber)
	if err != nil {
		return c.Status(fiber.StatusNotFound).SendString(fmt.Sprintf("Batch %d not found in local storage", batchNumber))
	}

	// Convert state root from decimal string to hex
	stateRootHex := proofData.NewStateRoot
	if val, ok := new(big.Int).SetString(proofData.NewStateRoot, 10); ok {
		// It's a decimal number, convert to hex
		stateRootHex = common.BigToHash(val).Hex()
	}

	resp := BatchResponse{
		BatchNumber:  proofData.BatchNumber,
		BatchHash:    proofData.BatchHash,
		NewStateRoot: stateRootHex,
		Submitter:    proofData.Submitter,
		Timestamp:    uint64(proofData.Timestamp),
		Status:       proofData.Status,
		TxHash:       proofData.TxHash,
		TxHashes:     proofData.TxHashes,
	}

	return c.JSON(resp)
}

// handleGetStatus godoc
// @Summary Get proof status
// @Description Get proof generation status for a batch
// @Tags Proofs
// @Param id path int true "Batch ID"
// @Produce json
// @Success 200 {object} StatusResponse
// @Failure 400 {string} string "Invalid batch ID"
// @Router /status/{id} [get]
func (api *APIServer) handleGetStatus(c *fiber.Ctx) error {
	idStr := c.Params("id")
	batchNumber, err := strconv.ParseUint(idStr, 10, 64)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).SendString("Invalid batch ID")
	}

	// Check ProverStore specifically for local proof status
	proof, err := api.store.GetProof(batchNumber)
	if err == nil && proof != nil {
		return c.JSON(StatusResponse{
			Status:      "COMPLETED",
			BatchNumber: batchNumber,
			Timestamp:   proof.Timestamp,
		})
	}

	// If not found in local store, check contract
	batchID := new(big.Int).SetUint64(batchNumber)
	batch, err := api.registry.GetBatch(&bind.CallOpts{}, batchID)
	if err == nil && batch.VerifiedAt.Uint64() > 0 {
		return c.JSON(StatusResponse{
			Status:      "VERIFIED_ON_CHAIN",
			BatchNumber: batchNumber,
			Note:        "Proof verified but local proof data not found",
		})
	}

	// Default to PENDING
	return c.JSON(StatusResponse{
		Status:      "PENDING",
		BatchNumber: batchNumber,
	})
}

// handleGetBatches godoc
// @Summary Get batches with pagination
// @Description Retrieve multiple batch metadata entries with pagination
// @Tags Batches
// @Param page query int false "Page number (default: 1)"
// @Param limit query int false "Items per page - 10, 25, 50, or 100 (default: 25)"
// @Produce json
// @Success 200 {object} map[string]interface{} "Paginated batch list"
// @Failure 400 {string} string "Invalid parameters"
// @Failure 500 {string} string "Server error"
// @Router /batches [get]
func (api *APIServer) handleGetBatches(c *fiber.Ctx) error {
	// Parse pagination parameters
	page := c.QueryInt("page", 1)
	limit := c.QueryInt("limit", 25)

	// Validate page
	if page < 1 {
		return c.Status(fiber.StatusBadRequest).SendString("Page must be >= 1")
	}

	// Validate and normalize limit
	validLimits := []int{10, 25, 50, 100}
	validLimit := false
	for _, vl := range validLimits {
		if limit == vl {
			validLimit = true
			break
		}
	}
	if !validLimit {
		// Default to closest valid limit
		if limit < 10 {
			limit = 10
		} else if limit < 25 {
			limit = 25
		} else if limit < 50 {
			limit = 50
		} else {
			limit = 100
		}
	}

	// Calculate offset
	offset := uint64((page - 1) * limit)

	// Fetch batches from storage
	batches, err := api.store.GetBatches(offset, uint64(limit))
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).SendString(fmt.Sprintf("Error fetching batches: %v", err))
	}

	// Get total count for pagination
	totalCount, err := api.store.GetTotalBatches()
	if err != nil {
		// Log error but continue with 0 count logic or handle gracefully
		fmt.Printf("Error fetching total batch count: %v\n", err)
	}

	// Calculate total pages
	totalPages := 0
	if limit > 0 {
		totalPages = int((totalCount + uint64(limit) - 1) / uint64(limit))
	}

	// Convert to BatchResponse objects
	response := make([]BatchResponse, len(batches))
	for i, b := range batches {
		// Convert state root from decimal string to hex if needed
		stateRootHex := b.NewStateRoot
		if val, ok := new(big.Int).SetString(b.NewStateRoot, 10); ok {
			// It's a decimal number, convert to hex
			stateRootHex = common.BigToHash(val).Hex()
		}

		response[i] = BatchResponse{
			BatchNumber:  b.BatchNumber,
			BatchHash:    b.BatchHash,
			TxHash:       b.TxHash,
			TxHashes:     b.TxHashes,
			Timestamp:    uint64(b.Timestamp),
			Submitter:    b.Submitter,
			NewStateRoot: stateRootHex,
			Status:       b.Status,
		}
	}

	// Return paginated response
	return c.JSON(fiber.Map{
		"page":        page,
		"limit":       limit,
		"count":       len(batches),
		"total_pages": totalPages,
		"batches":     response,
	})
}